move lowercase_keys into utils module, apply normalization to all requests
