+++ packages/kit/src/runtime/server/index.js
@@ -1,7 +1,8 @@
 import render_page from './page/index.js';
 import { render_response } from './page/render.js';
-import render_endpoint, { lowercase_keys } from './endpoint.js';
+import render_endpoint from './endpoint.js';
 import { parse_body } from './parse_body/index.js';
+import { lowercase_keys } from './utils.js';
 
 /**
  * @param {import('types/hooks').Incoming} incoming
@@ -20,12 +21,9 @@ export async function respond(incoming, options, state = {}) {
 		};
 	}
 
-	if (!incoming.headers['user-agent']) {
-		incoming.headers = lowercase_keys(incoming.headers);
-	}
-
 	const incoming_with_body = {
 		...incoming,
+		headers: lowercase_keys(incoming.headers),
 		body: parse_body(incoming)
 	};
